CTCP-3899
Updates to test navigation
